crawler_finalproject/requirements.txt
# 知乎热榜爬虫系统 - 依赖包列表
# Python版本: 3.10+

# ========== 爬虫相关 ==========
requests==2.31.0              # HTTP请求库
beautifulsoup4==4.12.2        # HTML解析库
lxml==4.9.3                   # XML/HTML解析器
scrapy==2.11.0                # 爬虫框架（可选）
selenium==4.15.2              # 浏览器自动化（应对动态加载）
playwright==1.40.0            # 现代化浏览器自动化（备选）

# ========== 数据存储 ==========
pymongo==4.6.0                # MongoDB驱动
mysql-connector-python==8.2.0 # MySQL驱动
redis==5.0.1                  # Redis驱动
sqlalchemy==2.0.23            # ORM框架

# ========== 数据处理与分析 ==========
pandas==2.1.4                 # 数据分析库
numpy==1.26.2                 # 数值计算库
openpyxl==3.1.2              # Excel文件处理

# ========== 数据可视化 ==========
matplotlib==3.8.2             # 绘图库
seaborn==0.13.0              # 统计图表库
plotly==5.18.0               # 交互式图表库
wordcloud==1.9.3             # 词云生成

# ========== Web框架（API服务） ==========
flask==3.0.0                  # 轻量级Web框架
fastapi==0.104.1             # 现代Web API框架
uvicorn==0.24.0              # ASGI服务器

# ========== 任务调度 ==========
apscheduler==3.10.4          # 任务调度库
celery==5.3.4                # 分布式任务队列
schedule==1.2.0              # 简单任务调度

# ========== 工具库 ==========
pyyaml==6.0.1                # YAML配置文件解析
python-dotenv==1.0.0         # 环境变量管理
loguru==0.7.2                # 日志库
tqdm==4.66.1                 # 进度条
fake-useragent==1.4.0        # 随机User-Agent
retrying==1.3.4              # 重试机制

# ========== 反爬虫相关 ==========
cloudscraper==1.2.71         # 绕过Cloudflare
httpx==0.25.2                # 现代HTTP客户端
pyppeteer==1.0.2             # Puppeteer Python版本

# ========== 验证码处理 ==========
pillow==10.1.0               # 图像处理
pytesseract==0.3.10          # OCR识别
ddddocr==1.4.11              # 验证码识别

# ========== 代理相关 ==========
aiohttp==3.9.1               # 异步HTTP客户端
proxy-pool==0.3.1            # 代理池

# ========== 测试相关 ==========
pytest==7.4.3                # 测试框架
pytest-cov==4.1.0            # 测试覆盖率
responses==0.24.1            # HTTP请求模拟

# ========== 类型检查与代码质量 ==========
mypy==1.7.1                  # 静态类型检查
black==23.12.0               # 代码格式化
flake8==6.1.0                # 代码风格检查
isort==5.13.2                # import排序

# ========== 其他依赖 ==========
cryptography==41.0.7         # 加密库
python-dateutil==2.8.2       # 日期处理
pytz==2023.3                 # 时区处理
click==8.1.7                 # 命令行接口